Add files via upload
this is my first programing challange. 
the objective is to understand and implement Object Oriented code. 
which was divided to three parts:

This project is concerned with the implementation of mathematical graphs and their usages. The basic definition of a graph is a set of nodes where some pairs of nodes are connected by edges, and where these edges have weights. Each edge of the graph has a direction so by default a graph is directional. 
part I:
Implement the Node class

part II:
Implement the Graph class

part III:
Implement the NonDirectionalGraph class as a sub-class of Graph.